# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema PetShop",
        "version": "2.02.3"
    },
    "servers": [
        {
            "url": "http://localhost/api"
        }
    ],
    "paths": {
...
        "/pet/cadastro": {
            "post": {
                "tags": [
                    "Pet"
                ],
                "operationId": "ebffda1a650c3e5de2ba1e198b1c5c0f",
                "parameters": [
                    {
                        "name": "name",
                        "required": true
                    },
                    {
                        "name": "peso",
                        "required": true
                    },
                    {
                        "name": "raca",
                        "required": true
                    },
                    {
                        "name": "sexo",
                        "required": true
                    },
                    {
                        "name": "user_id",
                        "required": true
                    },
                    {
                        "name": "especie",
                        "required": true
                    },
                    {
                        "name": "data_nascimento",
                        "required": true
                    },
                    {
                        "name": "data_falecimento",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cadastra as informações do Pet"
                    },
                    "401": {
                        "description": "Usuário não Autenticado"
                    },
                    "422": {
                        "description": "Erro em algum campo obrigatório"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/pet/editar/{id}": {
            "get": {
                "tags": [
                    "Pet"
                ],
                "operationId": "c9b50115ddfe3f1d6016d6045a1279f4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "/editar/{id}",
                        "description": "Pet id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Apresenta a informação do Pet selecionado"
                    },
                    "204": {
                        "description": "Pet não encotrado"
                    },
                    "401": {
                        "description": "Usuário não Autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/pet/atualizar/{id}": {
            "put": {
                "tags": [
                    "Pet"
                ],
                "operationId": "e6156b83407bf6a57622d8afd2ea738a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "/atualizar/{id}",
                        "description": "Pet id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "peso",
                        "required": true
                    },
                    {
                        "name": "raca",
                        "required": true
                    },
                    {
                        "name": "sexo",
                        "required": true
                    },
                    {
                        "name": "user_id",
                        "required": true
                    },
                    {
                        "name": "especie",
                        "required": true
                    },
                    {
                        "name": "data_nascimento",
                        "required": true
                    },
                    {
                        "name": "data_falecimento",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Atualiza as informações do Pet"
                    },
                    "401": {
                        "description": "Usuário não Autenticado"
                    },
                    "422": {
                        "description": "Erro em algum campo obrigatório"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/pet/listar": {
            "get": {
                "tags": [
                    "Pet"
                ],
                "operationId": "1f64c3dff0e45a1862de6157a3265083",
                "responses": {
                    "200": {
                        "description": "Lista os Pets cadastrados referentes ao usuário logado"
                    },
                    "401": {
                        "description": "Usuário não Autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/pet/select/{id}": {
            "get": {
                "tags": [
                    "Pet"
                ],
                "operationId": "e912ab4201db8c322d9752b0df9d77ae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "/select/{id}",
                        "description": "Pet id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Apreseta todos os Pets cadastrados referentes ao usuário logado"
                    },
                    "401": {
                        "description": "Usuário não Autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/procedimeto/cadastro": {
            "post": {
                "tags": [
                    "Procedimento"
                ],
                "operationId": "eae4bf5c3eb75358368ac8880716b2eb",
                "parameters": [
                    {
                        "name": "pet_id",
                        "required": true
                    },
                    {
                        "name": "vacina",
                        "required": true
                    },
                    {
                        "name": "castrado",
                        "required": true
                    },
                    {
                        "name": "cirurgia",
                        "required": true
                    },
                    {
                        "name": "banho_tosa",
                        "required": true
                    },
                    {
                        "name": "user_id",
                        "required": true
                    },
                    {
                        "name": "data_castracao",
                        "required": true
                    },
                    {
                        "name": "user_created",
                        "required": true
                    },
                    {
                        "name": "descricao_cirurgia",
                        "required": true
                    },
                    {
                        "name": "medicamento_id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cadastra as os procedimentos"
                    },
                    "401": {
                        "description": "Usuário não Autenticado"
                    },
                    "422": {
                        "description": "Erro em algum campo obrigatório"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/procedimeto/listar": {
            "get": {
                "tags": [
                    "Procedimento"
                ],
                "operationId": "100bf0648c2d99fe0055d0f9630abf1c",
                "responses": {
                    "200": {
                        "description": "Apresenta a informação do Pet selecionado"
                    },
                    "204": {
                        "description": "Pet não encotrado"
                    },
                    "401": {
                        "description": "Usuário não Autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}